Refactored all the game player objets to respond to the UUID change
